+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [20.2.1] 2020-02-18
+### Added
+- Added support for 'notEqual' in skip logic, to support '!=' in the 'enableWhen' of 
+  FHIR Questionnaire.
+
 ## [20.2.0] 2019-02-13
 ### Changed
 - Updated the fhirpath dependency to 1.0.1 (which was a breaking fhirpath